+++ packages/firebase_admin_sdk/lib/src/app.dart
@@ -35,7 +35,6 @@ import '../messaging.dart';
 import '../security_rules.dart';
 import '../storage.dart';
 import 'utils/utils.dart';
-import 'version.g.dart';
 
 part 'app/app_exception.dart';
 part 'app/app_options.dart';